fix: SDK7 raycast protocol update (#4761)
* Changed timestamp usage going back to having the same raycast's timestamp in the raycast result
* Implemented support for 3 new direction type (local direction, global direction, target entity, global target)
* Moved raycasting logic to a system instead of the RaycastHandler, to ensure:
    * Rays are casted after entity transforms resolving (using LateUpdate)
    * Being able to cope with the new `continuous` property that casts the ray on every frame regardless of having a model update
 * Created new 'SDKCustomLayer' in unity renderer and implemented sdk raycast component`collisionMask` property parsing into renderer layers (any `CL_CUSTOMX` => adds 'SDKCustomLayer' to the mask; `CL_POINTER` => 'OnPointerEvent'; `CL_PHYSICS` => 'CharacterController'; `CL_POINTER | CL_PHYSICS` => 'Default')
 * Updated `MeshCollider` collision layer handling in the renderer to take into account new `CL_CUSTOMX` layers (using 'SDKCustomLayer' in the renderer)
 * Updated raycast in the renderer to always use `RaycastAll` because of the case of different `CL_CUSTOM` layers all using the same 'SDKCustomLayer' in the renderer (and already existent CL_PHYSICS + CL_POINTER using the same 'Default' unity layer), and they are filtered out when the raycast result is constructed
 * Fixed layermask separate detection of CL_PHYSICS and CL_POINTER not being distinguished when setting only 1 of those in the raycast layermask

+++ unity-renderer/Assets/DCLPlugins/ECS7/Systems/UiPointerEvents/ECSUiPointerEventsSystem.cs
@@ -257,7 +257,7 @@ namespace ECSSystems.ECSUiPointerEventsSystem
                 type = pointerEventType,
                 hit = new RaycastHit
                 {
-                    EntityId = entity.entityId,
+                    EntityId = (uint)entity.entityId,
                     Length = 0,
                     Position = ProtoConvertUtils.UnityVectorToPBVector(position)
                 }
